extract layout and emoji group components
