+++ src/app/components/emoji-board/components/index.tsx
@@ -4,3 +4,5 @@ export * from './Sidebar';
 export * from './NoStickerPacks';
 export * from './Preview';
 export * from './Item';
+export * from './Group';
+export * from './Layout';
