+++ src/app/components/emoji-board/components/styles.css.ts
@@ -1,6 +1,31 @@
 import { style } from '@vanilla-extract/css';
 import { toRem, color, config, DefaultReset, FocusOutline } from 'folds';
 
+/**
+ * Layout
+ */
+
+export const Base = style({
+  maxWidth: toRem(432),
+  width: `calc(100vw - 2 * ${config.space.S400})`,
+  height: toRem(450),
+  backgroundColor: color.Surface.Container,
+  color: color.Surface.OnContainer,
+  border: `${config.borderWidth.B300} solid ${color.Surface.ContainerLine}`,
+  borderRadius: config.radii.R400,
+  boxShadow: config.shadow.E200,
+  overflow: 'hidden',
+});
+
+export const Header = style({
+  padding: config.space.S300,
+  paddingBottom: 0,
+});
+
+/**
+ * Sidebar
+ */
+
 export const Sidebar = style({
   width: toRem(54),
   backgroundColor: color.Surface.Container,
@@ -26,6 +51,10 @@ export const SidebarBtnImg = style({
   height: toRem(24),
   objectFit: 'contain',
 });
+
+/**
+ * Preview
+ */
 
 export const Preview = style({
   padding: config.space.S200,
@@ -55,6 +84,38 @@ export const PreviewImg = style([
     objectFit: 'contain',
   },
 ]);
+
+/**
+ * Group
+ */
+
+export const EmojiGroup = style({
+  position: 'relative',
+  padding: `${config.space.S300} 0`,
+});
+
+export const EmojiGroupLabel = style({
+  position: 'sticky',
+  top: config.space.S200,
+  zIndex: 1,
+
+  margin: 'auto',
+  padding: `${config.space.S100} ${config.space.S200}`,
+  borderRadius: config.radii.Pill,
+  backgroundColor: color.SurfaceVariant.Container,
+  color: color.SurfaceVariant.OnContainer,
+});
+
+export const EmojiGroupContent = style([
+  DefaultReset,
+  {
+    padding: `0 ${config.space.S200}`,
+  },
+]);
+
+/**
+ * Item
+ */
 
 export const EmojiItem = style([
   DefaultReset,
